fixed x axis position issue
